Refactor: Consolidate duplicate VolumeSlider components
Consolidates the two nearly identical VolumeSlider components into a single shared component to reduce code duplication and improve maintainability.

The new component at `components/shared/VolumeSlider.tsx` merges the functionality of the previous components, supporting props like `showValue` and `onVolumeChangeCommitted`. All call sites have been updated, and the old files have been deleted. A comprehensive test suite was created by merging the tests from both original components.
