+++ components/SpotifyDisplay.tsx
@@ -17,7 +17,7 @@ import IconButton from '@mui/material/IconButton'
 import Typography from '@mui/material/Typography'
 import { useCallback, useEffect, useReducer, useRef } from 'react'
 import AuthButton from './AuthButton'
-import VolumeSlider from './Spotify/VolumeSlider'
+import VolumeSlider from './shared/VolumeSlider'
 import SpotifyDeviceSelectorWrapper from './SpotifyDeviceSelectorWrapper'
 
 // 1. State Shape
@@ -447,6 +447,7 @@ const SpotifyDisplay = () => {
             onVolumeChange={handleVolumeChange}
             onVolumeChangeCommitted={handleVolumeChangeCommitted}
             onToggleMute={handleToggleMute}
+            showValue={true}
           />
           <SpotifyDeviceSelectorWrapper
             availableDevices={spotifyData.devices || []}
