+++ components/index.ts
@@ -19,7 +19,6 @@ export * from './TimerDisplay'
 export * from './TimerSoundProvider'
 export * from './WorkoutColumns'
 export * from './WorkoutTableViewer'
-export * from './PlaybackControls/VolumeSlider'
 export * from './Spotify/PlaylistSelector'
-export * from './Spotify/VolumeSlider'
 export * from './shared/StyledCard'
+export * from './shared/VolumeSlider'
