+++ tests/unit/app/client/control/components/SpotifyControls.test.tsx
@@ -98,7 +98,7 @@ describe('components/SpotifyControls', () => {
 
   it('should render the mute button with the correct aria-label', () => {
     render(<SpotifyControls />)
-    const muteButton = screen.getByLabelText(/mute volume/i)
+    const muteButton = screen.getByLabelText(/mute/i)
     expect(muteButton).toBeInTheDocument()
   })
